fixed bugs in question 4
didn't consider the effects of parameters of intersections that have been through on the stored answers. Sometimes because of the stored answers, the program may mistakenly think that there is no plans to get from one place to another. But the actual fact is that this is probably because during these searches that ended finding no plans at all, it found nothing because that some intersections are already included in the path